New Action: Facebook Like
Publish a Like on a Facebook item. Careful, though, as it requires an
optional app permission you probably want to check before giving the
user the ability to Like. Also gathering acting user's info and
permissions in one HTTP request and so more batch requests will be
needed in the future.